+++ Modules/Intent.Modules.Application.MediatR.Behaviours/NugetPackages.cs
@@ -10,6 +10,7 @@ namespace Intent.Modules.Application.MediatR.Behaviours
             name: "MediatR",
             version: outputTarget.GetMaxNetAppVersion() switch
             {
+                (6, 0) => "12.1.1",
                 _ => "12.4.0",
             });
 
@@ -17,7 +18,7 @@ namespace Intent.Modules.Application.MediatR.Behaviours
             name: "Microsoft.Extensions.Logging",
             version: outputTarget.GetMaxNetAppVersion() switch
             {
-                (6, 0) => "8.0.0",
+                (6, 0) => "6.0.0",
                 (7, 0) => "8.0.0",
                 _ => "8.0.0",
             });