+++ Modules/Intent.Modules.AspNetCore.MultiTenancy/Templates/MultiTenantStoreDbContext/MultiTenantStoreDbContextTemplatePartial.cs
@@ -22,9 +22,9 @@ namespace Intent.Modules.AspNetCore.MultiTenancy.Templates.MultiTenantStoreDbCon
         [IntentManaged(Mode.Ignore, Signature = Mode.Fully)]
         public MultiTenantStoreDbContextTemplate(IOutputTarget outputTarget, object model = null) : base(TemplateId, outputTarget, model)
         {
-            AddNugetDependency(NugetPackages.FinbuckleMultiTenantEntityFrameworkCore);
-            AddNugetDependency(NugetPackages.EntityFrameworkCoreInMemory(OutputTarget.GetProject()));
-            AddNugetDependency(NugetPackages.FinbuckleMultiTenantAspNetCore);
+            AddNugetDependency(NugetPackages.FinbuckleMultiTenantEntityFrameworkCore(outputTarget));
+            AddNugetDependency(NugetPackages.MicrosoftEntityFrameworkCoreInMemory(OutputTarget));
+            AddNugetDependency(NugetPackages.FinbuckleMultiTenantAspNetCore(outputTarget));
 
             CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
                 .AddUsing("Finbuckle.MultiTenant")