+++ Modules/Intent.Modules.Dapr.AspNetCore/NuGetPackages.cs
@@ -26,6 +26,7 @@ namespace Intent.Modules.Dapr.AspNetCore
             name: "MediatR",
             version: outputTarget.GetMaxNetAppVersion() switch
             {
+                (6, 0) => "12.1.1",
                 _ => "12.4.0",
             });