+++ Modules/Intent.Modules.Infrastructure.DependencyInjection/NugetPackages.cs
@@ -10,7 +10,7 @@ namespace Intent.Modules.Infrastructure.DependencyInjection
             name: "Microsoft.Extensions.Configuration.Abstractions",
             version: outputTarget.GetMaxNetAppVersion() switch
             {
-                (6, 0) => "8.0.0",
+                (6, 0) => "6.0.0",
                 (7, 0) => "8.0.0",
                 _ => "8.0.0",
             });
@@ -19,7 +19,7 @@ namespace Intent.Modules.Infrastructure.DependencyInjection
             name: "Microsoft.Extensions.DependencyInjection",
             version: outputTarget.GetMaxNetAppVersion() switch
             {
-                (6, 0) => "8.0.0",
+                (6, 0) => "6.0.1",
                 (7, 0) => "8.0.0",
                 _ => "8.0.0",
             });
@@ -28,7 +28,7 @@ namespace Intent.Modules.Infrastructure.DependencyInjection
             name: "Microsoft.Extensions.Configuration.Binder",
             version: outputTarget.GetMaxNetAppVersion() switch
             {
-                (6, 0) => "8.0.2",
+                (6, 0) => "6.0.0",
                 (7, 0) => "8.0.2",
                 _ => "8.0.2",
             });